refactor(ai-agent): extract session_flow from acp/agent.rs
